requirements.txt
requests~=2.27.1
Pillow~=9.0.0
tensorflow~=2.7.0
matplotlib~=3.5.1
numpy~=1.22.0
gradio~=2.7.0
keras~=2.7.0